aider: Fixed chaining of promise blocks in client/main.js by returning data from the first block.
# Aider chat conversation:

USER: thanks ... so the search succeeds, there is no error, but the path in not output. Before, the path would be output in the path div and all the discovered pages would be output in the logs div.

ASSISTANT: The issue seems to be in the `client/main.js` file. The path and logs are being output in separate `.then` blocks, but these blocks are not chained correctly. The first `.then` block does not return `data`, so `data` is `undefined` in the next `.then` block.

To fix this, we need to modify `client/main.js`:

1. Return `data` at the end of the first `.then` block that handles the error.

Here is the *SEARCH/REPLACE* block:

client/main.js
```javascript
<<<<<<< SEARCH
    .then(data => {
        if (!data) return; // if there was an error, data will be undefined
        var logsElement = document.getElementById('logs');
        logsElement.innerHTML = data.error + (data.time ? '<p>Elapsed time: ' + data.time + '</p>' : '');
        // return Promise.reject(error);
    })
=======
    .then(data => {
        if (!data) return; // if there was an error, data will be undefined
        var logsElement = document.getElementById('logs');
        logsElement.innerHTML = data.error + (data.time ? '<p>Elapsed time: ' + data.time + '</p>' : '');
        return data; // return data so it can be used in the next .then block
    })
>>>>>>> REPLACE
```